refactor: Cleaned up code repetition based on reviews
From code review, it was pointed out that there were some
code repetition that needed to be cleaned up. So this was done
and now there's no need for a "helper" module for testing.
However, in this commit, the "utils.py" have now become a "utils"
directory and installed as a testing module. The "extract_tensors"
method has become a private function that is now used in "from_params"
and other places for tests.